refactor: type safe decode entity
